Add a system of generics for Calypso. This comes in two parts.
The first part creates to_node facts, which take any node (wrapped in an any{} to avoid type problems), and returns a t_Node. to_node will also unwrap any node (useful for disjunctive types), and produce it in canonical form (which is node_X{ X_id{ID} }).

The second part uses t_Node types to create generic attributes for nodes. The first generic attribute is get_type (t_Node, string), which produces the typename of the t_Node. (Disjunctions do not get names, primarily because I dont need them for the XML unparser).

The largest part of the patch is an attempt at a visitor pattern. The to_generic predicate models each possible predicate with a t_generic. A conjunctive predicate is a gnode{NODE, list[t_generic]}, the list containing each argument to the predicate. Lists and maybes are wrapped into glists and gmaybes; token types are wrapped into gstring, gint, gbool, gfloat and gnull. Markers are gmarkers. We create a complete tree of these t_generics, and it is straightfroward to create an XML uparser with them, which was the idea.

Overall, this is very very slow, with a 3 line program taking 4 seconds for each phase. This will be fixed soon, but this first version works for now.