Refactoring, renaming and tidying up.
We no longer generate lots of predicates to ToPredName was no longer needed. Places that used it should have used 'contrsuctor' instead. Likewise, toDisjSubName was being used to get the type name as a String; this has been replaced with toTypeString. toDisjSubName was always used in conjunction with concatenating it with the toDisjBaseName, so they were combined. Some other duplicated code was merged into a single function.