toVarName has two purposes: to create variable names for fields and parameters, and to convert the name of the class to lower case for use in function names.
In the first case, however, we need to make a special case for C++ keywords like int, long etc. This adds an '_' in front of them (though the list is very incomplete).
In the second case we really dont need this '_'. So this creates a new set of functions toFuncPart, that are used everywhere that toVarName was used to create a function name.

We hadn't noticed this before in phc, as classes like INT, BOOL etc were never used directly - always through an Expr or a Literal, so there was never a case where a class had a field called 'int'. That has changed in the LIR (though, I suppose I could have worked around it by renaming the field).

Tested by checking that the AST, HIR and MIR are unchanged, and that underscores are appropriately added to the LIR.